Split plutus scripts into separate library.
